Removida altiga CurrencyApi
-- Com o rework da curencyApi completo, as arquivos antigos foram removidos e o nome das pastas atualizado. A currencyApi agora é feita em typescript
-- A classe currencyApi foi movida para um arquivo currencyApi.ts, o export da instanciação continuou no index. Isso faz com que, quando compilado, a instância da currencyApi seja feita usando module.exports ao invés de exports.default
-- Modificado nome de alguns métodos da currencyApi para deixar claro (para um import usando CommonJS) que é um método privado
-- Métodos privados de execução única, como inicializadores (listener, new_transaction) foram renomeados para começar com 2 underlines
-- Modificados algumas configurações de compilação
-- Adicionada pasta build na .gitignore
