+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
+build/
 .idea/*
 .vscode/*
 !.vscode/settings.json
