+++ package.json
@@ -11,8 +11,7 @@
     "eslint": "eslint .",
     "eslint-fix": "eslint --fix .",
     "testdaemon": "env-cmd -f ./cfg/test.env jest --watchAll --verbose",
-    "dev-server": "env-cmd -f ./cfg/dev.env nodemon --inspect src/index.js",
-    "dev-server-ts": "env-cmd -f ./cfg/dev.env ts-node-dev src/index.ts --inspect",
+    "dev-server": "env-cmd -f ./cfg/dev.env ts-node-dev src/index.js --inspect",
     "dev-client": "npm start --prefix ./client"
   },
   "jest": {
@@ -29,24 +28,24 @@
     "nano-address-validator": "^3.0.0",
     "randomstring": "^1.1.5",
     "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.0",
     "socket.io-stream": "^0.9.1"
   },
   "devDependencies": {
     "@types/express": "^4.17.2",
     "@types/node": "^12.12.7",
     "@types/socket.io": "^2.1.4",
+    "@typescript-eslint/eslint-plugin": "^2.7.0",
+    "@typescript-eslint/parser": "^2.7.0",
     "concurrently": "^5.0.0",
     "cookieparser": "^0.1.0",
     "env-cmd": "^10.0.1",
     "eslint": "^6.6.0",
-    "eslint-plugin-flowtype": "^4.3.0",
+    "eslint-plugin-flowtype": "^4.4.1",
     "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.16.0",
     "eslint-plugin-react-hooks": "^2.2.0",
     "jest": "^24.8.0",
-    "nodemon": "^1.19.4",
     "supertest": "^4.0.2",
     "ts-node-dev": "^1.0.0-pre.44",
     "typescript": "^3.7.2"
