--- src/currencyApi/currencies/bitcoin.js
@@ -1,9 +1,0 @@
-/**
- * src/api/internal/bitcoin.js
- */
-
-module.exports = {
-	code: 'btc',
-	ip: process.env.BITCOIN_IP || 'localhost',
-	port: 8091,
-}
