+++ src/currencyApi/currencies/common/methods/connection.ts
