+++ src/currencyApi/currencies/common/methods/garbage_collector.ts
