+++ src/currencyApi/currencies/common/methods/index.ts
