--- src/currencyApi/currencies/nano.js
@@ -1,9 +1,0 @@
-/**
- * src/api/internal/nano.js
- */
-
-module.exports = {
-	code: 'nano',
-	ip: process.env.NANO_IP || 'localhost',
-	port: 8090,
-}
