--- src/currencyApi/self/listener/functions/get/index.js
@@ -1,19 +1,0 @@
-/*
- * src/currencyApi/self/listener/functions/get/index.js
- * 
- * Esse módulo exporta um objeto com todos os módulos dessa pasta, os módulos
- * são acessíveis em uma propriedade com o nome do módulo. Ex.: this.modulo irá
- * retornar o retorno de um require('modulo.js') contido nessa pasta
- */
-
-const normalizedPath = require('path').join(__dirname)
-
-const get_functions = {}
-
-require('fs').readdirSync(normalizedPath)
-	.forEach(filename =>
-		filename !== 'index.js' &&
-	(get_functions[filename.replace('.js', '')] = require(`./${filename}`))
-	)
-
-module.exports = get_functions
