--- src/currencyApi/self/listener/functions/get/ping.js
@@ -1,9 +1,0 @@
-/*
- * src/currencyApi/self/listener/functions/get/ping.js
- * 
- * Retorna pong
- */
-
-module.exports = function ping() {
-	return 'pong'
-}
